end goal, npc added; pathfollow made more generic
Path following script made more generic so that it can be reused for the NPC characters as well as trucks. Truck controls created. Win and lose conditions are more fleshed out. Collision added to the truck for NPC hit interaction. Camera follow has a slight delay for a smoother follow with lerp.